+++ bootique-shiro-web-mdc/src/test/java/io/bootique/shiro/web/mdc/ShiroWebModuleProviderTest.java
@@ -1,0 +1,12 @@
+package io.bootique.shiro.web.mdc;
+
+import io.bootique.test.junit.BQModuleProviderChecker;
+import org.junit.Test;
+
+public class ShiroWebModuleProviderTest {
+
+    @Test
+    public void testPresentInJar() {
+        BQModuleProviderChecker.testPresentInJar(ShiroWebMDCModuleProvider.class);
+    }
+}